+++ RELEASE_NOTES.md
@@ -13,6 +13,7 @@
 + Add support for Syslog RemoteServers
 + Add schema reference documentation
 + DeviceGroup.owner is now required
++ Fix bug in which DO sets task status to ERROR right away while it is still rolling back
 + When targetSshKey is used try bash shell to modify targetUsername password if tmsh shell fails.
 
 ## Version 1.6.0